+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -19,6 +19,7 @@
 #include "Rule12p3Check.h"
 #include "Rule14p4Check.h"
 #include "Rule18p7Check.h"
+#include "Rule18p8Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -48,6 +49,8 @@ public:
         "misrac-rule-14p4");
     CheckFactories.registerCheck<Rule18p7Check>(
         "misrac-rule-18p7");
+    CheckFactories.registerCheck<Rule18p8Check>(
+        "misrac-rule-18p8");
   }
 };